requirements.txt
python-telegram-bot==21.0.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
Flask==3.0.0
